+++ src/statistics/distributions/bernoulli.rs
@@ -4,6 +4,8 @@
 // See LICENSE or <https://www.gnu.org/licenses/>.
 // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
+use crate::statistics::DistributionError;
+
 use super::Distribution;
 use num_complex::Complex;
 
@@ -105,7 +107,7 @@ impl Distribution for Bernoulli {
         1.0 - self.p + self.p * f64::exp(t)
     }
 
-    fn sample(&self, n: usize) -> Vec<f64> {
+    fn sample(&self, n: usize) -> Result<Vec<f64>, DistributionError> {
         // IMPORT HERE TO AVOID CLASH WITH
         // `RustQuant::distributions::Distribution`
         use rand::thread_rng;
@@ -115,7 +117,7 @@ impl Distribution for Bernoulli {
 
         let mut rng = thread_rng();
 
-        let dist = Bernoulli::new(self.p).unwrap();
+        let dist = Bernoulli::new(self.p)?;
 
         let mut variates: Vec<f64> = Vec::with_capacity(n);
 
@@ -123,7 +125,7 @@ impl Distribution for Bernoulli {
             variates.push(dist.sample(&mut rng) as usize as f64);
         }
 
-        variates
+        Ok(variates)
     }
 }
 
@@ -273,16 +275,18 @@ mod tests_bernoulli {
     #[should_panic]
     fn test_sample_zero_size() {
         let bernoulli = Bernoulli::new(0.5);
-        bernoulli.sample(0);
+        _ = bernoulli.sample(0);
     }
 
     #[test]
-    fn test_sample_positive_size() {
+    fn test_sample_positive_size() -> Result<(), DistributionError> {
         let bernoulli = Bernoulli::new(0.5);
-        let sample = bernoulli.sample(100);
+        let sample = bernoulli.sample(100)?;
         assert_eq!(sample.len(), 100);
         for &value in sample.iter() {
             assert!(value == 0.0 || value == 1.0);
         }
+
+        Ok(())
     }
 }