+++ src/statistics/distributions/gamma.rs
@@ -4,7 +4,7 @@
 // See LICENSE or <https://www.gnu.org/licenses/>.
 // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-use crate::statistics::distributions::Distribution;
+use crate::statistics::{distributions::Distribution, DistributionError};
 use num_complex::Complex;
 use statrs::function::gamma::{gamma, gamma_li};
 
@@ -115,7 +115,7 @@ impl Distribution for Gamma {
         (1.0 - t / self.beta).powf(-self.alpha)
     }
 
-    fn sample(&self, n: usize) -> Vec<f64> {
+    fn sample(&self, n: usize) -> Result<Vec<f64>, DistributionError> {
         // IMPORT HERE TO AVOID CLASH WITH
         // `RustQuant::distributions::Distribution`
         use rand::thread_rng;
@@ -125,7 +125,7 @@ impl Distribution for Gamma {
 
         let mut rng = thread_rng();
 
-        let dist = Gamma::new(self.alpha, self.beta.recip()).unwrap();
+        let dist = Gamma::new(self.alpha, self.beta.recip())?;
 
         let mut variates: Vec<f64> = Vec::with_capacity(n);
 
@@ -133,7 +133,7 @@ impl Distribution for Gamma {
             variates.push(dist.sample(&mut rng) as usize as f64);
         }
 
-        variates
+        Ok(variates)
     }
 }